requirements.txt
openai>=1.0.0
anthropic>=0.8.0
requests>=2.25.0
schedule>=1.2.0
python-dotenv>=0.19.0
pyyaml>=6.0
pytest>=7.0
